+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `texlab.hover.symbols` setting to allow turning off image preview for symbol-like commands ([#1420](https://github.com/latex-lsp/texlab/issues/1420))
+
 ## [5.23.1] - 2025-06-24
 
 ### Fixed